added pvalue
Former-commit-id: 2b65d6c42d07f5406ee9a9e7d018d1313f935d6f

+++ go/src/spexs/pools.go
@@ -40,7 +40,7 @@ func (p *FifoPool) Len() int {
 	return p.list.Len()
 }
 
-type TrieFitnessFunc func(p *TrieNode) float32
+type TrieFitnessFunc func(p *TrieNode) float64
 
 type PriorityPool struct {
 	token   chan int
